Ensure "OP_0" and "false" have the same represenstation on stack
During the script execution via `executeScript()` certain operations
add boolean values to the stack.

Boolean values, which are added to the stack as result, are represented
as follows in Bitcoin Core:

- `true`: as vector with one element: `[1]`
- `false`: as empty vector with zero elements: `[]`

See:

- https://github.com/bitcoin/bitcoin/blob/48efbdbe986355bd2478f0fdd366b20952fbf30a/src/script/interpreter.cpp#L244
- https://github.com/bitcoin/bitcoin/blob/48efbdbe986355bd2478f0fdd366b20952fbf30a/src/script/interpreter.cpp#L246

However, in BitcoinJ the representation of `false` differs in some
cases, where it is represented as Byte array with one element, which
is zero: `[0]`

See:

- `OP_EQUAL`: https://github.com/bitcoinj/bitcoinj/blob/5dcf6439755ac5db6e151ba4403b5d3e8bf792eb/core/src/main/java/org/bitcoinj/script/Script.java#L1025
- `OP_CHECKSIG`: https://github.com/bitcoinj/bitcoinj/blob/5dcf6439755ac5db6e151ba4403b5d3e8bf792eb/core/src/main/java/org/bitcoinj/script/Script.java#L1304
- `OP_CHECKMULTISIG`: https://github.com/bitcoinj/bitcoinj/blob/5dcf6439755ac5db6e151ba4403b5d3e8bf792eb/core/src/main/java/org/bitcoinj/script/Script.java#L1381

At minimum this has an impact on the behavior of `OP_EQUAL` and
`OP_EQUALVERIFY`, when comparing something with `OP_FALSE`/`OP_0`.

This commit attemps to fix the issue, to mirror the verification
behavior of Bitcoin Core, by adding empty Byte arrays to the stack,
instead of Byte arrays with `0`.